refactor(cli-engine): handlers are direct function references (ruling R-A)
Operator ruling: do not dynamically import handlers. The lazy-handler
pattern (handler: () => Promise<{ default: Handler }>) is abolished for
result, session, and server commands; `handler` is now the handler
function itself. The engine drops the module-loading step and its error
handling; the whoami port inlines its handler into whoami.ts (the
handler file is deleted), which also removes the circular type-inference
friction recorded in the divergence doc. CommandHandler<D> remains for
statically imported handlers. Draft \S6 amended accordingly
(operator-authorized).
