+++ packages/cli-engine/tests/events.test.ts
@@ -14,71 +14,69 @@ const T0 = "1970-01-01T00:00:00.000Z";
 
 const noisy = defineCommand({
   help: { summary: "Emits the whole vocabulary" },
-  handler: async () => ({
-    default: async (_args, ctx) => {
-      ctx.report({ kind: "step-started", step: "compile", id: "s1" });
-      ctx.report({
-        kind: "progress",
-        step: "compile",
-        completed: 1,
-        total: 2,
-      });
-      ctx.report({
-        kind: "step-finished",
-        step: "compile",
-        id: "s1",
-        outcome: "ok",
-      });
-      ctx.report({ kind: "step-finished", step: "lint", outcome: "warning" });
-      ctx.report({ kind: "message", severity: "warn", text: "heads up" });
-      ctx.report({ kind: "message", severity: "info", text: "fyi" });
-      ctx.report({
-        kind: "output",
-        source: "generator",
-        channel: "data",
-        line: "generated 3 files",
-      });
-      ctx.report({
-        kind: "output",
-        source: "generator",
-        channel: "diagnostic",
-        line: "generator warmed up",
-      });
-      ctx.report({
-        kind: "remediation",
-        action: { kind: "run-command", label: "Review", command: "demo show" },
-      });
-      ctx.report({
-        kind: "endpoint",
-        name: "studio",
-        url: "http://localhost:5555",
-      });
-      ctx.report({
-        kind: "status",
-        subject: "db",
-        status: "ready",
-        from: "starting",
-      });
-      ctx.report({
-        kind: "artifact",
-        path: "out/contract.json",
-        description: "the contract",
-        data: { bytes: 42 },
-      });
-      return ok(
-        ctx.present(
-          { data: { done: true } },
-          {
-            human: (): readonly Block[] => [
-              { kind: "summary", tone: "ok", text: "done" },
-            ],
-            stdout: () => ["done"],
-            next: () => [{ kind: "done", label: "Nothing else" }],
-          },
-        ),
-      );
-    },
-  }),
+  handler: async (_args, ctx) => {
+    ctx.report({ kind: "step-started", step: "compile", id: "s1" });
+    ctx.report({
+      kind: "progress",
+      step: "compile",
+      completed: 1,
+      total: 2,
+    });
+    ctx.report({
+      kind: "step-finished",
+      step: "compile",
+      id: "s1",
+      outcome: "ok",
+    });
+    ctx.report({ kind: "step-finished", step: "lint", outcome: "warning" });
+    ctx.report({ kind: "message", severity: "warn", text: "heads up" });
+    ctx.report({ kind: "message", severity: "info", text: "fyi" });
+    ctx.report({
+      kind: "output",
+      source: "generator",
+      channel: "data",
+      line: "generated 3 files",
+    });
+    ctx.report({
+      kind: "output",
+      source: "generator",
+      channel: "diagnostic",
+      line: "generator warmed up",
+    });
+    ctx.report({
+      kind: "remediation",
+      action: { kind: "run-command", label: "Review", command: "demo show" },
+    });
+    ctx.report({
+      kind: "endpoint",
+      name: "studio",
+      url: "http://localhost:5555",
+    });
+    ctx.report({
+      kind: "status",
+      subject: "db",
+      status: "ready",
+      from: "starting",
+    });
+    ctx.report({
+      kind: "artifact",
+      path: "out/contract.json",
+      description: "the contract",
+      data: { bytes: 42 },
+    });
+    return ok(
+      ctx.present(
+        { data: { done: true } },
+        {
+          human: (): readonly Block[] => [
+            { kind: "summary", tone: "ok", text: "done" },
+          ],
+          stdout: () => ["done"],
+          next: () => [{ kind: "done", label: "Nothing else" }],
+        },
+      ),
+    );
+  },
 });
 
 function makeCli() {
