--- packages/cli/src/v8/auth/whoami.handler.ts
@@ -1,102 +1,0 @@
-import type { CommandHandler, Presentations } from "@prisma/cli-engine";
-import {
-  CliStructuredError,
-  type NextAction,
-  notOk,
-  ok,
-} from "@prisma/cli-engine/protocol";
-import { EmptyServiceTokenError, readAuthState } from "../../lib/auth/auth-ops";
-import type { AuthProviderId, AuthStateResult } from "../../types/auth";
-import type { authWhoamiCommand } from "./whoami";
-
-const TITLE = "Showing the current authenticated identity.";
-
-const SIGN_IN: NextAction = {
-  kind: "run-command",
-  label: "Sign in",
-  command: "prisma-cli auth login",
-};
-
-interface FieldRow {
-  readonly label: string;
-  readonly value: string;
-}
-
-function providerLabel(provider: AuthProviderId): string {
-  return provider === "github" ? "GitHub" : "Google";
-}
-
-function userLabel(state: AuthStateResult): string | null {
-  if (state.user?.email) {
-    return state.user.email;
-  }
-
-  if (state.credential?.type === "service_token") {
-    return state.credential.name
-      ? `<service token: ${state.credential.name}>`
-      : "<service token>";
-  }
-
-  if (state.credential?.type === "management_token") {
-    return state.credential.name
-      ? `<management token: ${state.credential.name}>`
-      : "<management token>";
-  }
-
-  return null;
-}
-
-function fieldRows(state: AuthStateResult): readonly FieldRow[] {
-  if (!state.authenticated) {
-    return [{ label: "status", value: "signed out" }];
-  }
-
-  const user = userLabel(state);
-  return [
-    { label: "status", value: "signed in" },
-    ...(user ? [{ label: "user", value: user }] : []),
-    ...(state.provider
-      ? [{ label: "provider", value: providerLabel(state.provider) }]
-      : []),
-    ...(state.workspace?.name
-      ? [{ label: "workspace", value: state.workspace.name }]
-      : []),
-  ];
-}
-
-function presentationsFor(state: AuthStateResult): Presentations {
-  const rows = fieldRows(state);
-  return {
-    human: () => [
-      { kind: "summary", tone: "info", text: TITLE },
-      { kind: "fields", rows },
-    ],
-    stdout: () => rows.map((row) => `${row.label}: ${row.value}`),
-    next: () => (state.authenticated ? [] : [SIGN_IN]),
-  };
-}
-
-const run: CommandHandler<typeof authWhoamiCommand> = async (_args, ctx) => {
-  let state: AuthStateResult;
-  try {
-    state = await readAuthState(process.env, ctx.signal);
-  } catch (error) {
-    if (error instanceof EmptyServiceTokenError) {
-      return notOk(
-        new CliStructuredError(
-          "AUTH.CONFIG_INVALID",
-          "Authentication configuration is invalid",
-          {
-            why: error.message,
-            fix: "Provide a valid PRISMA_SERVICE_TOKEN value, or unset the variable to use local OAuth login.",
-          },
-        ),
-      );
-    }
-    throw error;
-  }
-
-  return ok(ctx.present({ data: state }, presentationsFor(state)));
-};
-
-export default run;
